requirements.txt
streamlit
google-genai
pandas
pillow
openpyxl
streamlit-drawable-canvas
pymupdf
msal
